+++ tests/Jroc.Tests/Function/JavaScript/Function_NewExpression_ForOfReassignedDeclaration_DispatchesReplacement.js
@@ -1,0 +1,12 @@
+function Point(x) {
+    this.value = x + 1;
+}
+
+for ({ Point } of [{
+    Point: function (x) {
+        this.value = x + 20;
+    }
+}]) {
+}
+
+console.log(new Point(5).value);