+++ tests/Jroc.Tests/Function/JavaScript/Function_NewExpression_ReassignedDeclaration_DispatchesReplacement.js
@@ -1,0 +1,9 @@
+function Point(x) {
+    this.value = x + 1;
+}
+
+Point = function (x) {
+    this.value = x + 10;
+};
+
+console.log(new Point(5).value);